# Linear implementation for Fulkerson-Chen-Anstee