--- a/include/igraph_graphicality.h
+++ b/include/igraph_graphicality.h
@@ -24,7 +24,6 @@
 #include "igraph_decls.h"
 #include "igraph_error.h"
 #include "igraph_vector.h"
-#include "igraph_vector_list.h"
 
 __BEGIN_DECLS