Closes #1452

The current implementation of the Fulkerson-Chen-Anstee simple directed graphicality check is quadratic.
This is my own approach to the linear time version of Fulkerson-Chen-Anstee.
I (with @szhorvat) plan to do a write-up for this approach.

After implementing the method I read section 5.1 in:
H. Kim, C. I. Del Genio, K. E. Bassler, and Z. Toroczkai, New J. Phys. 14, 023012 (2012). https://dx.doi.org/10.1088/1367-2630/14/2/023012
The methods follow the same general approach.
The key difference is they use recurrence relations to prove that the method works.
I leave here an explanation of how my method works and as in the other methods, the first step is sorting.

The theorem (screenshotted from Wikipedia):

<img width="755" alt="image" src="https://github.com/igraph/igraph/assets/1212871/f919397d-f23b-4f23-90c9-a2a52d3a1999">


The left hand side is easy to compute in linear time for each value of k.

On the right hand side we have two sums of outdegrees.
For the first sum we separate outdegrees based on whether they are $< k-1$ or $>= k-1$.
Similarly, for the second sum we separate outdegrees based on whether they are $< k$ or $>= k$.
In both cases we have a value which we can think of as the splitting point of a sum.
We only need to track the sum of outdegrees below the splitting point, because the value of k is incrementing from $1$ to $n$, and therefore they will not individually be impacted by future changes.

The outdegrees that are equal to or above the splitting point must be tracked, and we do so with an ordered multiset.
The ordered set for the first sum is initially empty and the one for the second sum initially contains all of the outdegrees.
As $k$ increases, either add/remove the outdegree at index k-1 to/from the sum or ordered set, depending on which side of the splitting point it lies.
We may also need to remove the smallest outdegrees from the ordered sets and move them over to the sums of outdegrees below the splitting point.
The contribution of the values in the ordered multiset to the right hand side is the value of $k-1$ or $k$, depending on which sum we are considering, multiplied by the size of the ordered multiset.

Using a balanced binary search tree for the ordered set means adding and removing from the set takes $\mathcal{O}(\log n)$ time.
Then the time complexity will match that of the sort operation.

We can push this a little further.
Since the values in the set are bounded by n, we can switch the quicksort out for a bucket sort which is $\mathcal{O}(n)$.
Additionally, it allows us to switch out a balanced binary search tree for an array of integers, where the index in to the array represents the degree and the value at that index represents how many times that degree appears in the ordered multiset.
All operations within the main loop are $\mathcal{O}(1)$ and therefore the time complexity of the algorithm is $\mathcal{O}(n)$ in total.
I have not done any benchmarking yet, but I have tested an equivalent C++ implementation on sequences of size $1$ to $2\,000\,000$ and there is a noticeable difference in running time between the linear and the linearithmic approaches.