@szhorvat 2024-03-22
I edited the theorem into the top post so it's easier to follow your description, and what the LHS and RHS refer to.