@szhorvat 2024-03-31
I added some rudimentary benchmarks.

----

Before this PR:

```
|> Benchmark file: /Users/szhorvat/Repos/igraph-main/igraph/tests/benchmarks/graphicality.c
| Undirected simple,       PA, n = 50, 2000000x                                     0.27s   0.27s      0s
| Undirected loops,        PA, n = 50, 2000000x                                    0.635s  0.633s      0s
| Undirected multi,        PA, n = 50, 2000000x                                    0.055s  0.055s      0s
| Undirected multi, loops, PA, n = 50, 2000000x                                        0s      0s      0s
| Undirected simple,       G(n,p), n = 50, 2000000x                                  0.2s  0.198s      0s
| Undirected loops,        G(n,p), n = 50, 2000000x                                0.789s  0.786s      0s
| Undirected multi,        G(n,p), n = 50, 2000000x                                0.053s  0.053s      0s
| Undirected multi, loops, G(n,p), n = 50, 2000000x                                    0s      0s      0s
| Directed simple,       PA, n = 50, 2000000x                                       1.26s   1.26s      0s
| Directed loops,        PA, n = 50, 2000000x                                       1.18s   1.17s      0s
| Directed multi,        PA, n = 50, 2000000x                                      0.061s   0.06s      0s
| Directed multi, loops, PA, n = 50, 2000000x                                          0s      0s      0s
| Directed simple,       G(n,p), n = 50, 2000000x                                   1.24s   1.23s      0s
| Directed loops,        G(n,p), n = 50, 2000000x                                   1.17s   1.17s      0s
| Directed multi,        G(n,p), n = 50, 2000000x                                  0.061s   0.06s      0s
| Directed multi, loops, G(n,p), n = 50, 2000000x                                      0s      0s      0s

| Undirected simple,       PA, n = 1000, 100000x                                    0.13s  0.128s      0s
| Undirected loops,        PA, n = 1000, 100000x                                   0.544s   0.54s      0s
| Undirected multi,        PA, n = 1000, 100000x                                   0.059s  0.058s      0s
| Undirected multi, loops, PA, n = 1000, 100000x                                       0s      0s      0s
| Undirected simple,       G(n,p), n = 1000, 100000x                               0.101s    0.1s      0s
| Undirected loops,        G(n,p), n = 1000, 100000x                               0.659s  0.654s      0s
| Undirected multi,        G(n,p), n = 1000, 100000x                               0.059s  0.058s      0s
| Undirected multi, loops, G(n,p), n = 1000, 100000x                                   0s      0s      0s
| Directed simple,       PA, n = 1000, 100000x                                      1.83s   1.82s      0s
| Directed loops,        PA, n = 1000, 100000x                                     0.825s  0.821s      0s
| Directed multi,        PA, n = 1000, 100000x                                      0.06s  0.059s      0s
| Directed multi, loops, PA, n = 1000, 100000x                                         0s      0s      0s
| Directed simple,       G(n,p), n = 1000, 100000x                                   1.8s    1.8s      0s
| Directed loops,        G(n,p), n = 1000, 100000x                                 0.827s  0.825s      0s
| Directed multi,        G(n,p), n = 1000, 100000x                                  0.06s  0.059s      0s
| Directed multi, loops, G(n,p), n = 1000, 100000x                                     0s      0s      0s

| Undirected simple,       PA, n = 1000000, 100x                                   0.152s  0.134s  0.017s
| Undirected loops,        PA, n = 1000000, 100x                                   0.995s  0.968s  0.021s
| Undirected multi,        PA, n = 1000000, 100x                                   0.059s  0.058s      0s
| Undirected multi, loops, PA, n = 1000000, 100x                                   0.058s  0.057s      0s
| Undirected simple,       G(n,p), n = 1000000, 100x                                0.12s  0.102s  0.017s
| Undirected loops,        G(n,p), n = 1000000, 100x                                1.42s   1.39s  0.022s
| Undirected multi,        G(n,p), n = 1000000, 100x                               0.059s  0.058s      0s
| Undirected multi, loops, G(n,p), n = 1000000, 100x                               0.058s  0.057s      0s
| Directed simple,       PA, n = 1000000, 100x                                      19.5s   19.4s  0.028s
| Directed loops,        PA, n = 1000000, 100x                                      1.47s   1.42s  0.042s
| Directed multi,        PA, n = 1000000, 100x                                     0.062s  0.061s      0s
| Directed multi, loops, PA, n = 1000000, 100x                                     0.059s  0.058s      0s
| Directed simple,       G(n,p), n = 1000000, 100x                                  19.6s   19.6s  0.028s
| Directed loops,        G(n,p), n = 1000000, 100x                                  1.47s   1.42s  0.043s
| Directed multi,        G(n,p), n = 1000000, 100x                                 0.061s   0.06s      0s
| Directed multi, loops, G(n,p), n = 1000000, 100x                                 0.059s  0.058s      0s
```

----

After this PR:

```
|> Benchmark file: /Users/szhorvat/Repos/igraph-main/igraph/tests/benchmarks/graphicality.c
| Undirected simple,       PA, n = 50, 2000000x                                    0.251s   0.25s      0s
| Undirected loops,        PA, n = 50, 2000000x                                      0.6s  0.598s      0s
| Undirected multi,        PA, n = 50, 2000000x                                    0.053s  0.053s      0s
| Undirected multi, loops, PA, n = 50, 2000000x                                        0s      0s      0s
| Undirected simple,       G(n,p), n = 50, 2000000x                                0.185s  0.183s      0s
| Undirected loops,        G(n,p), n = 50, 2000000x                                0.767s  0.764s      0s
| Undirected multi,        G(n,p), n = 50, 2000000x                                0.054s  0.053s      0s
| Undirected multi, loops, G(n,p), n = 50, 2000000x                                    0s      0s      0s
| Directed simple,       PA, n = 50, 2000000x                                       3.06s   3.05s      0s
| Directed loops,        PA, n = 50, 2000000x                                       1.13s   1.13s      0s
| Directed multi,        PA, n = 50, 2000000x                                      0.061s   0.06s      0s
| Directed multi, loops, PA, n = 50, 2000000x                                          0s      0s      0s
| Directed simple,       G(n,p), n = 50, 2000000x                                   3.06s   3.05s      0s
| Directed loops,        G(n,p), n = 50, 2000000x                                   1.14s   1.13s      0s
| Directed multi,        G(n,p), n = 50, 2000000x                                   0.06s   0.06s      0s
| Directed multi, loops, G(n,p), n = 50, 2000000x                                      0s      0s      0s

| Undirected simple,       PA, n = 1000, 100000x                                    0.13s  0.128s      0s
| Undirected loops,        PA, n = 1000, 100000x                                   0.542s  0.538s      0s
| Undirected multi,        PA, n = 1000, 100000x                                   0.059s  0.058s      0s
| Undirected multi, loops, PA, n = 1000, 100000x                                       0s      0s      0s
| Undirected simple,       G(n,p), n = 1000, 100000x                               0.101s    0.1s      0s
| Undirected loops,        G(n,p), n = 1000, 100000x                               0.656s  0.654s      0s
| Undirected multi,        G(n,p), n = 1000, 100000x                               0.059s  0.058s      0s
| Undirected multi, loops, G(n,p), n = 1000, 100000x                                   0s      0s      0s
| Directed simple,       PA, n = 1000, 100000x                                      2.23s   2.23s      0s
| Directed loops,        PA, n = 1000, 100000x                                     0.803s  0.799s      0s
| Directed multi,        PA, n = 1000, 100000x                                     0.061s  0.059s      0s
| Directed multi, loops, PA, n = 1000, 100000x                                         0s      0s      0s
| Directed simple,       G(n,p), n = 1000, 100000x                                  2.24s   2.23s      0s
| Directed loops,        G(n,p), n = 1000, 100000x                                 0.804s  0.801s      0s
| Directed multi,        G(n,p), n = 1000, 100000x                                  0.06s  0.059s      0s
| Directed multi, loops, G(n,p), n = 1000, 100000x                                     0s      0s      0s

| Undirected simple,       PA, n = 1000000, 100x                                   0.151s  0.133s  0.017s
| Undirected loops,        PA, n = 1000000, 100x                                   0.991s  0.966s  0.021s
| Undirected multi,        PA, n = 1000000, 100x                                   0.059s  0.058s      0s
| Undirected multi, loops, PA, n = 1000000, 100x                                   0.058s  0.057s      0s
| Undirected simple,       G(n,p), n = 1000000, 100x                                0.12s  0.103s  0.017s
| Undirected loops,        G(n,p), n = 1000000, 100x                                1.42s   1.39s  0.022s
| Undirected multi,        G(n,p), n = 1000000, 100x                               0.059s  0.058s      0s
| Undirected multi, loops, G(n,p), n = 1000000, 100x                               0.059s  0.057s      0s
| Directed simple,       PA, n = 1000000, 100x                                      2.54s   2.16s  0.357s
| Directed loops,        PA, n = 1000000, 100x                                      1.45s   1.39s  0.048s
| Directed multi,        PA, n = 1000000, 100x                                      0.06s   0.06s      0s
| Directed multi, loops, PA, n = 1000000, 100x                                     0.059s  0.058s      0s
| Directed simple,       G(n,p), n = 1000000, 100x                                  2.56s   2.16s   0.37s
| Directed loops,        G(n,p), n = 1000000, 100x                                  1.46s   1.39s  0.049s
| Directed multi,        G(n,p), n = 1000000, 100x                                 0.061s  0.061s      0s
| Directed multi, loops, G(n,p), n = 1000000, 100x                                  0.06s  0.059s      0s
```